--- src/main/kotlin/containsDuplicate/ContainsDuplicate.kt
@@ -1,12 +1,0 @@
-package containsDuplicate
-
-class ContainsDuplicate {
-    fun containsDuplicate(nums: IntArray): Boolean {
-        val set = mutableSetOf<Int>()
-        for (num in nums) {
-            if (set.contains(num)) return true
-            set.add(num)
-        }
-        return false
-    }
-}
