+++ src/main/kotlin/me/bossm0n5t3r/leetcode/containsDuplicate/ContainsDuplicate.kt
@@ -1,0 +1,14 @@
+package me.bossm0n5t3r.leetcode.containsDuplicate
+
+class ContainsDuplicate {
+    class Solution {
+        fun containsDuplicate(nums: IntArray): Boolean {
+            val set = mutableSetOf<Int>()
+            for (num in nums) {
+                if (set.contains(num)) return true
+                set.add(num)
+            }
+            return false
+        }
+    }
+}
